requirements.txt
pinecone_haystack
haystack-ai
fastapi
uvicorn
python-dotenv
pathlib
sentence-transformers
python-multipart
-e .